Update Pharma_data_py.py
Pandas, Seaborn, and SKlearn packages are used to explore the Pharma dataset. We first explore basic statistics to get an idea of the distribution of the data, and then develop both a linear regression model and random forest model to predict whether prescriptions are from a specialist or a general practitioner.